requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.7
pillow==11.0.0
requests==2.31.0
setuptools==69.0.3
termcolor==2.3.0
typing_extensions==4.9.0
urllib3==2.1.0
yaspin==3.1.0
